Refactor MDX components and update imports
- Removed unused imports from mdx-components.js to streamline the code.
- Updated the import path for defaultComponents in Editor.tsx to reference the new client-mdx-components file.
- Enhanced the TSDoc usage in mdx-components.js for improved documentation clarity.
