+++ packages/docs/mdx-components.js
@@ -1,23 +1,13 @@
 import { useMDXComponents as getDocsMDXComponents } from 'nextra-theme-docs'
-import { Application, Entity } from '@playcanvas/react'
-import { OrbitControls } from '@playcanvas/react/scripts'
-import { Align, Light, Anim, Camera, Collision, EnvAtlas, GSplat, Script, Render, RigidBody } from '@playcanvas/react/components'
 
-import ReactQueryProvider from '@/docs-components/ReactQueryProvider'
-
-import EnvAtlasComponent from '@components/EnvAtlas'
-import Grid from '@components/Grid'
-import PostEffects, { StaticPostEffects} from '@components/PostEffects'
-import ShadowCatcher from '@components/ShadowCatcher'
-import AutoRotate from '@components/AutoRotate'
-import { MotionEntity, MotionLight } from '@components/MotionEntity'
-import { Glb } from '@components/Glb'
 import { unstable_TSDoc as TSDoc } from 'nextra/tsdoc'
+import { defaultComponents } from './client-mdx-components';
+
 import { version } from 'playcanvas';
 const PcVersion = () => <span>{version}</span>;
 
 const docsComponents = getDocsMDXComponents({
-  CustomTSDoc(props) {
+  TSDoc(props) {
     return (
       <TSDoc
         {...props}
@@ -31,24 +21,6 @@ const docsComponents = getDocsMDXComponents({
     )
   }
 })
-
-export const defaultComponents = {
-  PcVersion,
-  ReactQueryProvider,
-  Application,
-  Entity,
-  Align, Anim, Camera, Collision, EnvAtlas, GSplat, Script, Render, RigidBody, Light,
-  OrbitControls,
-  EnvAtlasComponent,
-  Grid,
-  ShadowCatcher,
-  PostEffects,
-  StaticPostEffects,
-  AutoRotate,
-  MotionEntity,
-  MotionLight,
-  Glb
-}
 
 export function useMDXComponents(components) {
   return {
